+++ src/main/java/team/washer/server/v2/domain/user/service/SearchUserService.java
@@ -2,6 +2,6 @@ package team.washer.server.v2.domain.user.service;
 
 import team.washer.server.v2.domain.user.dto.UserListResponseDto;
 
-public interface QueryUsersByFilterService {
+public interface SearchUserService {
     UserListResponseDto getUsersByFilter(String name, String roomNumber, Integer grade, Integer floor);
 }
